김준영 : enum 및 HomeWork 시작.

+++ CPlusPlus/GameEngineConsole/GameEngineArray.h
@@ -66,7 +66,7 @@ public:
 		return ArrPtr[_Index];
 	}
 
-	void ReSize(int _Value) 
+	void ReSize(size_t _Value) 
 	{
 		// 20줄 안팍.
 		// 삼항 연산자 써보시면 좋을겁니다.
@@ -77,7 +77,7 @@ public:
 		// 기존의 있던 값에서 현재의 배열이 복사한다음 삭제해야 한다.
 
 		DataType* NewPtr = new DataType[_Value];
-		int CopySize = _Value < ArrCount ? _Value : ArrCount;
+		size_t CopySize = _Value < ArrCount ? _Value : ArrCount;
 
 		for (size_t i = 0; i < CopySize; i++)
 		{
